refactor: split drive_tools_testable.go into domain files (Closes aliwatters/gsuite-mcp#107)
Split the 1136-line monolithic drive_tools_testable.go into four domain-specific files:
- drive_files_testable.go: file CRUD, download/upload (TestableDriveSearch through TestableDriveDelete)
- drive_sharing_testable.go: permissions and sharing (TestableDriveShare, TestableDriveGetPermissions, TestableDriveGetShareableLink)
- drive_comments_testable.go: comments and replies (TestableDriveListComments through TestableDriveCreateReply)
- drive_revisions_testable.go: revision operations (TestableDriveListRevisions through TestableDriveDownloadRevision)
- drive_shared_helpers.go: shared utilities (extractRequiredFileID, readLimited, downloadFileContent)

All tests pass. No behavior changes.
